refactor (#611)
